requirements.txt
bokeh
can4python
numpy
pandas
PyYAML
scipy
click